Added a Jet Enemy
The Jet enemy spawns at the top of the screen and is removed when it reaches the bottom.

despawnOffScreenObjects was modified for objects that reach the left and right border of the screen (this is for a different type of jet to be added that will move from left to right)

Bullet collision and spawning for the Jet still need some work.